+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+## [Unreleased]
+### Added
+- Shell completion scripts for kctl-env (Bash and Zsh)
+
 ## [v0.1.0] - 2026-01-07
 ### Added
 - Initial scaffold: bin/kubectl shim with env/.kubectl-version/global/auto resolution and TTL cache
@@ -15,3 +19,5 @@ All notable changes to this project will be documented in this file. This projec
 - Git repository initialized, branch protections configured (squash/rebase, linear history, delete branch on merge)
 
 [v0.1.0]: https://github.com/senet/kctl-env/releases/tag/v0.1.0
+
+[Unreleased]: https://github.com/senet/kctl-env/compare/v0.1.0...HEAD